requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
jinja2==3.1.4

celery==5.3.6
redis==5.0.1
ccxt==4.2.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
